chore: Move proof and policy packages under new witness package
Move proof and policy packages under new witness package. This folder will also contain selection package that will select witnesses based on policy.

Closes #863
