+++ cmd/orb-server/startcmd/start.go
@@ -86,8 +86,8 @@ import (
 	"github.com/trustbloc/orb/pkg/anchor/handler/credential"
 	"github.com/trustbloc/orb/pkg/anchor/handler/proof"
 	"github.com/trustbloc/orb/pkg/anchor/linkstore"
-	"github.com/trustbloc/orb/pkg/anchor/policy"
-	policyhandler "github.com/trustbloc/orb/pkg/anchor/policy/resthandler"
+	"github.com/trustbloc/orb/pkg/anchor/witness/policy"
+	policyhandler "github.com/trustbloc/orb/pkg/anchor/witness/policy/resthandler"
 	"github.com/trustbloc/orb/pkg/anchor/writer"
 	"github.com/trustbloc/orb/pkg/cas/extendedcasclient"
 	ipfscas "github.com/trustbloc/orb/pkg/cas/ipfs"
