+++ pkg/anchor/witness/policy/resthandler/configurator_test.go
